CTCTRADERS-327: move header and rename attributes
